Assessing probability of success

Assessment of the probability of truly or falsely (depending on simulated scenario) rejecting the null hypothesis of interest for a given weight and evidence level, using simulated data as input.

Usage

oc_pos(
  m,
  se,
  probs,
  weights = seq(0, 1, by = 0.01),
  map_prior,
  sigma,
  null_effect = 0,
  direction_pos = T,
  n_cores = 1,
  eval_strategy = "sequential"
)

Arguments

m

Numerical vector of simulated effect estimates.

se

Numerical vector of simulated standard errors (m and se need to have the same length).

probs

Vector of quantiles q, with 1 minus q representing an evidence level of interest (where positive effect estimate indicate a beneficial treatment).

weights

Vector of weights of the informative component of the MAP prior (defaults to seq(0, 1, by = 0.01)).

map_prior

A MAP prior containing information about the trials in the source population, created using RBesT; a mixture of normal distributions is required.

sigma

Standard deviation of the weakly informative component of the MAP prior, recommended to be the unit-information standard deviation.

null_effect

Numerical value, representing the null effect (defaults to 0).

direction_pos

Logical value, TRUE (default) if effects greater that the null_effect indicate a beneficial treatment and FALSE otherwise.

n_cores

Integer value, representing the number of cores to be used (defaults to 1); only applies if eval_strategy is not "sequential".

eval_strategy

Character variable, representing the evaluation strategy, either "sequential", "multisession", or "multicore" (see documentation of future::plan, defaults to "sequential").

Value

A 2-dimensional array containing probabilities, either of truly (probability of success) or falsely rejecting the null hypothesis of interest for a given weight and evidence level.

See also

Examples

set.seed(123)
n_sims <- 5 # small number for exemplary application
sim_dat <- list(
  "m" = rnorm(n = n_sims, mean = 1.15, sd = 0.1),
  "se" = rnorm(n = n_sims, mean = 1.8, sd = 0.3)
)
results <- oc_pos(
  m = sim_dat[["m"]],
  se = sim_dat[["se"]],
  probs = c(0.025, 0.05, 0.1, 0.2), 
  weights = seq(0, 1, by = 0.01), 
  map_prior = load_tipmap_data("tipmapPrior.rds"), 
  sigma = 16.23,
  null_effect = 0,
  direction_pos = TRUE, 
  eval_strategy = "sequential",
  n_cores = 1
) 
print(results)
#>        q=0.025 q=0.05 q=0.1 q=0.2
#> w=0        0.0    0.0   0.0   0.2
#> w=0.01     0.0    0.0   0.0   0.2
#> w=0.02     0.0    0.0   0.0   0.4
#> w=0.03     0.0    0.0   0.0   0.6
#> w=0.04     0.0    0.0   0.0   0.6
#> w=0.05     0.0    0.0   0.0   0.6
#> w=0.06     0.0    0.0   0.0   0.8
#> w=0.07     0.0    0.0   0.0   0.8
#> w=0.08     0.0    0.0   0.2   0.8
#> w=0.09     0.0    0.0   0.2   1.0
#> w=0.1      0.0    0.0   0.2   1.0
#> w=0.11     0.0    0.0   0.2   1.0
#> w=0.12     0.0    0.0   0.2   1.0
#> w=0.13     0.0    0.0   0.2   1.0
#> w=0.14     0.0    0.0   0.2   1.0
#> w=0.15     0.0    0.0   0.4   1.0
#> w=0.16     0.0    0.0   0.6   1.0
#> w=0.17     0.0    0.0   0.6   1.0
#> w=0.18     0.0    0.0   0.6   1.0
#> w=0.19     0.0    0.0   0.6   1.0
#> w=0.2      0.0    0.0   0.6   1.0
#> w=0.21     0.0    0.0   0.6   1.0
#> w=0.22     0.0    0.0   0.8   1.0
#> w=0.23     0.0    0.0   0.8   1.0
#> w=0.24     0.0    0.2   0.8   1.0
#> w=0.25     0.0    0.2   0.8   1.0
#> w=0.26     0.0    0.2   0.8   1.0
#> w=0.27     0.0    0.2   0.8   1.0
#> w=0.28     0.0    0.2   1.0   1.0
#> w=0.29     0.0    0.2   1.0   1.0
#> w=0.3      0.0    0.2   1.0   1.0
#> w=0.31     0.0    0.2   1.0   1.0
#> w=0.32     0.0    0.2   1.0   1.0
#> w=0.33     0.0    0.2   1.0   1.0
#> w=0.34     0.0    0.2   1.0   1.0
#> w=0.35     0.0    0.4   1.0   1.0
#> w=0.36     0.0    0.4   1.0   1.0
#> w=0.37     0.0    0.6   1.0   1.0
#> w=0.38     0.0    0.6   1.0   1.0
#> w=0.39     0.0    0.6   1.0   1.0
#> w=0.4      0.0    0.6   1.0   1.0
#> w=0.41     0.0    0.6   1.0   1.0
#> w=0.42     0.0    0.6   1.0   1.0
#> w=0.43     0.0    0.6   1.0   1.0
#> w=0.44     0.0    0.6   1.0   1.0
#> w=0.45     0.0    0.6   1.0   1.0
#> w=0.46     0.0    0.8   1.0   1.0
#> w=0.47     0.0    0.8   1.0   1.0
#> w=0.48     0.0    0.8   1.0   1.0
#> w=0.49     0.0    0.8   1.0   1.0
#> w=0.5      0.2    0.8   1.0   1.0
#> w=0.51     0.2    0.8   1.0   1.0
#> w=0.52     0.2    0.8   1.0   1.0
#> w=0.53     0.2    0.8   1.0   1.0
#> w=0.54     0.2    1.0   1.0   1.0
#> w=0.55     0.2    1.0   1.0   1.0
#> w=0.56     0.2    1.0   1.0   1.0
#> w=0.57     0.2    1.0   1.0   1.0
#> w=0.58     0.2    1.0   1.0   1.0
#> w=0.59     0.2    1.0   1.0   1.0
#> w=0.6      0.2    1.0   1.0   1.0
#> w=0.61     0.2    1.0   1.0   1.0
#> w=0.62     0.2    1.0   1.0   1.0
#> w=0.63     0.2    1.0   1.0   1.0
#> w=0.64     0.4    1.0   1.0   1.0
#> w=0.65     0.4    1.0   1.0   1.0
#> w=0.66     0.4    1.0   1.0   1.0
#> w=0.67     0.6    1.0   1.0   1.0
#> w=0.68     0.6    1.0   1.0   1.0
#> w=0.69     0.6    1.0   1.0   1.0
#> w=0.7      0.6    1.0   1.0   1.0
#> w=0.71     0.6    1.0   1.0   1.0
#> w=0.72     0.6    1.0   1.0   1.0
#> w=0.73     0.6    1.0   1.0   1.0
#> w=0.74     0.6    1.0   1.0   1.0
#> w=0.75     0.8    1.0   1.0   1.0
#> w=0.76     0.8    1.0   1.0   1.0
#> w=0.77     0.8    1.0   1.0   1.0
#> w=0.78     0.8    1.0   1.0   1.0
#> w=0.79     0.8    1.0   1.0   1.0
#> w=0.8      0.8    1.0   1.0   1.0
#> w=0.81     0.8    1.0   1.0   1.0
#> w=0.82     1.0    1.0   1.0   1.0
#> w=0.83     1.0    1.0   1.0   1.0
#> w=0.84     1.0    1.0   1.0   1.0
#> w=0.85     1.0    1.0   1.0   1.0
#> w=0.86     1.0    1.0   1.0   1.0
#> w=0.87     1.0    1.0   1.0   1.0
#> w=0.88     1.0    1.0   1.0   1.0
#> w=0.89     1.0    1.0   1.0   1.0
#> w=0.9      1.0    1.0   1.0   1.0
#> w=0.91     1.0    1.0   1.0   1.0
#> w=0.92     1.0    1.0   1.0   1.0
#> w=0.93     1.0    1.0   1.0   1.0
#> w=0.94     1.0    1.0   1.0   1.0
#> w=0.95     1.0    1.0   1.0   1.0
#> w=0.96     1.0    1.0   1.0   1.0
#> w=0.97     1.0    1.0   1.0   1.0
#> w=0.98     1.0    1.0   1.0   1.0
#> w=0.99     1.0    1.0   1.0   1.0
#> w=1        1.0    1.0   1.0   1.0